chore: Correct placement of withFailMessage call
Resolve Sonar warnings of type java:S5833 - AssertJ methods setting the assertion context should come before an assertion

related to #88

+++ engine/src/main/java/org/operaton/bpm/engine/impl/metrics/MetricsQueryImpl.java
@@ -36,7 +36,7 @@ import org.operaton.bpm.engine.management.MetricIntervalValue;
 public class MetricsQueryImpl extends ListQueryParameterObject implements Serializable, Command<Object>, MetricsQuery {
 
   public static final int DEFAULT_LIMIT_SELECT_INTERVAL = 200;
-  public static final int DEFAULT_SELECT_INTERVAL = 15 * 60;
+  public static final long DEFAULT_SELECT_INTERVAL = 15 * 60L;
 
   @Serial
   private static final long serialVersionUID = 1L;
